requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
langgraph>=0.1.5
langchain-core>=0.3.2
requests==2.32.3
fastapi==0.115.2
uvicorn==0.30.3
